Created the working API
Created all of the curl-scripts to help with testing if the application
is working correctly.

created product_routes file to begin routing requests and responses
server side.

made the create route with .post to ensure that a user can make post a
product.

Registered the routes file to be used under the server.js file

successfully made the create, index, update and destroy routes without
throwing errors.